Refactor webapp routes into dedicated routers
